Add notes and examples on how to remove an element from normal flow with floats.
When you float an item, it is removed from the normal flow and pulled to the far left of the container it is in. If the floated element has elements below it, they will move up against each other as if the floated element isn't in the document. If you want to push elements that are in normal flow away from a floated element, you would add margin to the floated element, not the elements in normal flow.